\name{kernels}
\alias{biweight}
\alias{epanechnikov}
\alias{rectangular}
\alias{triangular}

\title{
Univariate Kernels
}

\description{
Evaluate univariate kernels.
}

\usage{
biweight(x)
epanechnikov(x)
rectangular(x)
triangular(x)
}


\arguments{
  \item{x}{a numeic vector.}
}

\value{
a numeric vector.
}

\author{
Kjell Konis \email{kjell.konis@me.com}
}

\keyword{multivariate}

