\name{methods.mvkde}
\alias{print.mvkde}
\alias{summary.mvkde}
\alias{print.summary.mvkde}

\title{
mvkde methods
}

\description{
Print and summary methods for mvkde objects.
}

\usage{
  \method{print}{mvkde}(x, ...)
  \method{summary}{mvkde}(object, ...)
  \method{print}{summary.mvkde}(x, ...)
}

\arguments{
  \item{x}{an \code{mvkde} or \code{summary.mvkde} object.}
  \item{object}{an mvkde object.}
  \item{\dots}{addtional arguments.}
}

\author{
Kjell Konis \email{kjell.konis@me.com}
}

\keyword{methods}


