\name{WandJones}
\alias{Uncorrelated.Normal}
\alias{Correlated.Normal}
\alias{Skewed}
\alias{Kurtotic}
\alias{Bimodal1}
\alias{Bimodal2}
\alias{Bimodal3}
\alias{Bimodal4}
\alias{Trimodal1}
\alias{Trimodal2}
\alias{Trimodal3}
\alias{Quadrimodal}

\docType{data}

\title{
Example Data Sets from Wand & Jones
}

\description{
The 12 normal mixture densities considered in Wand and Jones (1993).
}

\usage{
data(Uncorrelated.Normal)
data(Correlated.Normal)
data(Skewed)
data(Kurtotic)
data(Bimodal1)
data(Bimodal2)
data(Bimodal3)
data(Bimodal4)
data(Trimodal1)
data(Trimodal2)
data(Trimodal3)
data(Quadrimodal)
}

\details{
These 12 data sets consist of 500 points drawn from the Gaussian mixture distributions described in Wand and Jones (1993). The code is provided in the \code{scripts} subdirectory of this package.
}

\source{
Wand, M. P. and Jones, M. C. (1993). Comparison of Smoothing Parameterizations in Bivariate Kernel Density Estimation. \emph{Journal of the American Statistical Association}, Vol. 88, No. 422, pp. 520-528.
}

\keyword{datasets}
